refactor imms table
