+++ packages/web/app/features/ImmunisationsTable/index.js
@@ -1,0 +1,1 @@
+export * from './ImmunisationsTable';
