+++ packages/web/app/features/index.js
@@ -1,0 +1,1 @@
+export * from './ImmunisationsTable';
